feat: add View/Layout/Surface typed wrappers and fix javagen for string/float constants
- Fix javagen constants template to handle string and float32 types
  (View class has AUTOFILL_HINT_* strings and REQUESTED_FRAME_RATE_* floats)
- Convert Java float suffix (NaNf, -4.0f) to Go syntax in specgen
- Handle non-const-compatible values (NaN) as var declarations
- Add ref classes: SurfaceView, SurfaceHolder, Surface, View, ViewGroup,
  FrameLayout, Resources
- Regenerate all specs and Go wrappers

+++ app/alarm/consts/consts.go
@@ -16,9 +16,9 @@ const (
 )
 
 const (
-	IntervalDay            int = 86400000
-	IntervalFifteenMinutes int = 900000
-	IntervalHalfDay        int = 43200000
-	IntervalHalfHour       int = 1800000
-	IntervalHour           int = 3600000
+	IntervalDay            int64 = 86400000
+	IntervalFifteenMinutes int64 = 900000
+	IntervalHalfDay        int64 = 43200000
+	IntervalHalfHour       int64 = 1800000
+	IntervalHour           int64 = 3600000
 )
